fix: adjust pr changes
