+++ src/email/formatters.ts
@@ -1,0 +1,5 @@
+import { isEmail } from './validators'
+
+export const getEmailDomain = (arg: string): string | undefined => {
+  return isEmail(arg) ? arg.toLowerCase().split('@').pop() : ''
+}
