+++ src/email/index.ts
@@ -1,3 +1,4 @@
 export * from './awsSES'
+export * from './formatters'
 export * from './templates'
 export * from './validators'
